Correct import pathing
So the mistake here is that once your in the util folder you can load all the other files using just import <name> because you can always know their going to be top level and together, it's only when your in a different folder that you need to prefix with the full path. 

Or at least that's my understanding, this is one of the python things I don't get perfectly and should probably read up on.