Update timeseries data interface
This commit updates the interface for timerseries metadata to
be at the browbeat run level. the get_timeseries() generator
will provide all timeseries data for a Browbeat run iff
the named parameter timeseries is True on browbeat run declaration.

Note that on declaration graphite will be hit and trawled for all
the different metric names, I don't suggest toggling this option
to true unless you intend to use the graphite data. The rest of the
data is in an interator only to keep local memory requirements down.